Move getResources to utility.
