+++ packages/analysis-utils/src/browserManagement/getResources.ts
